requirements.txt
jsonschema
Pillow
numpy
